Factor out error helpers into a new cli module
